Record restore timing for async restores
