+++ corehq/apps/ota/views.py
@@ -1,7 +1,5 @@
-import logging
 from distutils.version import LooseVersion
 
-from django.conf import settings
 from django.http import JsonResponse
 from django.urls import reverse
 from django.shortcuts import redirect
@@ -38,42 +36,15 @@ from corehq.apps.ota.tasks import queue_prime_restore
 from corehq.apps.users.models import CommCareUser, CouchUser
 from corehq.apps.locations.permissions import location_safe
 from corehq.form_processor.exceptions import CaseNotFound
-from corehq.util.datadog.gauges import datadog_counter
 from dimagi.utils.decorators.memoized import memoized
 from casexml.apps.phone.restore import RestoreConfig, RestoreParams, RestoreCacheSettings
+from casexml.apps.phone.utils import record_restore_timing
 from django.http import HttpResponse
 from soil import MultipleTaskDownload
 
 from .utils import (
     demo_user_restore_response, get_restore_user, is_permitted_to_restore,
     handle_401_response, update_device_id)
-
-
-RESTORE_SEGMENTS = {
-    "FixtureElementProvider": "fixtures",
-    "CasePayloadProvider": "cases",
-}
-
-
-def _get_time_bucket(duration):
-    """Get time bucket for the given duration
-
-    Bucket restore times because datadog's histogram is too limited
-
-    Basically restore frequency is not high enough to have a meaningful
-    time distribution with datadog's 10s aggregation window, especially
-    with tags. More details:
-    https://help.datadoghq.com/hc/en-us/articles/211545826
-    """
-    if duration < 5:
-        return "lt_005s"
-    if duration < 20:
-        return "lt_020s"
-    if duration < 60:
-        return "lt_060s"
-    if duration < 120:
-        return "lt_120s"
-    return "over_120s"
 
 
 @location_safe
@@ -87,23 +58,6 @@ def restore(request, domain, app_id=None):
     """
     response, timing_context = get_restore_response(
         domain, request.couch_user, app_id, **get_restore_params(request))
-    tags = [
-        u'status_code:{}'.format(response.status_code),
-    ]
-    env = settings.SERVER_ENVIRONMENT
-    if (env, domain) in settings.RESTORE_TIMING_DOMAINS:
-        tags.append(u'domain:{}'.format(domain))
-    if timing_context is not None:
-        for timer in timing_context.to_list(exclude_root=True):
-            if timer.name in RESTORE_SEGMENTS:
-                segment = RESTORE_SEGMENTS[timer.name]
-                bucket = _get_time_bucket(timer.duration)
-                datadog_counter(
-                    'commcare.restores.{}'.format(segment),
-                    tags=tags + ['duration:%s' % bucket],
-                )
-        tags.append('duration:%s' % _get_time_bucket(timing_context.duration))
-    datadog_counter('commcare.restores.count', tags=tags)
     return response
 
 
@@ -277,15 +231,9 @@ def get_restore_response(domain, couch_user, app_id=None, since=None, version='1
         case_sync=case_sync,
     )
     response = restore_config.get_response()
-    timing = restore_config.timing_context
-    if timing.duration > 20 or response.status_code == 412:
-        sync_log_id = None
-        if restore_config.restore_state.current_sync_log:
-            sync_log_id = restore_config.restore_state.current_sync_log._id
-        log = logging.getLogger(__name__)
-        log.info("restore %s: domain=%s status=%s duration=%.3f",
-                 sync_log_id, domain, response.status_code, timing.duration)
-    return response, timing
+    if not async_restore_enabled:
+        record_restore_timing(restore_config, response.status_code)
+    return response, restore_config.timing_context
 
 
 class PrimeRestoreCacheView(BaseSectionPageView, DomainViewMixin):
